refactor(utils): moved clean_up_keywords to dqm_ingest specific utils
