+++ agr_literature_service/lit_processing/data_ingest/dqm_ingest/parse_dqm_json_reference.py
@@ -10,8 +10,8 @@ from os import environ, makedirs, path
 import bs4
 from dotenv import load_dotenv
 
-from agr_literature_service.lit_processing.utils.file_processing_utils import (clean_up_keywords, split_identifier,
-                                                                               write_json)
+from agr_literature_service.lit_processing.data_ingest.dqm_ingest.utils.dqm_processing_utils import clean_up_keywords
+from agr_literature_service.lit_processing.utils.file_processing_utils import split_identifier, write_json
 from agr_literature_service.lit_processing.utils.tmp_files_utils import init_tmp_dir
 
 init_tmp_dir()
